+++ src/main/java/com/hkbu/comp7640/entity/User.java
@@ -23,13 +23,13 @@ public class User implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "user_id", type = IdType.AUTO)
-    private Long userId;
+    private Integer userId;
 
     private String userName;
 
     private String nickName;
 
-    private Long contactNumber;
+    private Integer contactNumber;
 
     private String shippingDetails;
