+++ src/main/java/com/hkbu/comp7640/service/impl/TransactionServiceImpl.java
@@ -25,7 +25,7 @@ public class TransactionServiceImpl extends ServiceImpl<TransactionMapper, Trans
     private TransactionMapper transactionMapper;
 
     @Override
-    public IPage<TransactionWithProductVendorDTO> pageTransaction(Page<TransactionWithProductVendorDTO> page, Long userId) {
+    public IPage<TransactionWithProductVendorDTO> pageTransaction(Page<TransactionWithProductVendorDTO> page, Integer userId) {
         return transactionMapper.pageTransaction(page, userId);
     }
